refactor: new auth_core app
