+++ backend/backend/apps/core_auth/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CoreAuthConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "backend.apps.core_auth"
