+++ backend/backend/apps/core_auth/models.py
@@ -1,0 +1,2 @@
+
+# Create your models here.
